+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/ExtraClassesKitTest.groovy
@@ -1,0 +1,44 @@
+package ru.vyarus.gradle.plugin.animalsniffer
+
+import org.gradle.testkit.runner.BuildResult
+import org.gradle.testkit.runner.TaskOutcome
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 19.03.2017
+ */
+class ExtraClassesKitTest extends AbstractKitTest {
+
+    def "Check ignored classes"() {
+        setup:
+        build """
+            plugins {
+                id 'java'
+                id 'ru.vyarus.animalsniffer'
+            }
+
+            animalsniffer {
+                ignoreFailures = true
+                ignore 'java.nio.file.Paths'
+            }
+
+            repositories { mavenCentral()}
+            dependencies {
+                signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
+            }
+
+        """
+        fileFromClasspath('src/main/java/invalid/Sample.java', '/ru/vyarus/gradle/plugin/animalsniffer/java/invalid/Sample.java')
+//        debug()
+
+        when: "run task"
+        BuildResult result = run('check')
+
+        then: "task successful"
+        result.task(':check').outcome == TaskOutcome.SUCCESS
+
+        then: "found 1 violation instead of 2"
+        result.output.contains("1 AnimalSniffer violations were found in 1 files")
+        result.output.contains("int Boolean.compare(boolean, boolean)")
+    }
+}