+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Change console reporting format (to match gradle-quality-plugin format):
     - Each error split to source and code lines followed by empty line
     - IDE will be able to show link for class reference (clickable) when line number available
+* Add ignored classes configuration: extra allowed classes not present in signature
 
 ### 1.2.0 (2016-08-27)
 * Gradle 3.0 compatibility